Fix insert for faq and activities
