+++ src/components/Activities/Activities.jsx
@@ -1,17 +1,16 @@
 import React, { Component } from 'react';
 import {render} from 'react-dom';
-import Grid from 'react-bootstrap/lib/Grid'
-import Row from 'react-bootstrap/lib/Row'
 import { BootstrapTable, TableHeaderColumn } from 'react-bootstrap-table';
+import DateTime from 'react-datetime';
+import moment from 'moment';
 import {
-  getAllQuestions,
-  handleQuestionEdit,
-  handleDeleteQuestions,
-  handleInsertQuestion
-} from './FAQ_utils';
+  getAllActivities,
+  handleEditActivities,
+  handleDeleteActivities,
+  handleInsertActivities,
+} from './Activities_utils';
 
-
-class FAQ extends Component {
+class Activities extends Component {
   constructor(props) {
     super(props);
     this.setTableData = this.setTableData.bind(this);
@@ -20,31 +19,14 @@ class FAQ extends Component {
     this.beforeSaveCell = this.beforeSaveCell.bind(this);
 
     this.state = {
-      tableData: [
-        {
-          id: 0,
-          question: "Where are the washrooms?",
-          answer: "The washrooms are located by the museum. Please consult the map for directions."
-        },
-        {
-          id: 1,
-          question: "How does parking work?",
-          answer: "The parking is located by the entrance. There will be ushers there to guide you."
-        },
-        {
-          id: 2,
-          question: "What should I bring?",
-          answer: "Sunscreen, water, and a lunch. A towel to sit on during lunch may also come in handy, as lunch will be held outside."
-
-        }
-      ]
+      tableData: []
     };
   }
 
   componentDidMount() {
-    var questionsPromise = getAllQuestions();
-    questionsPromise.then(response => {
-      if(response.data){
+    let activitiesPromise = getAllActivities();
+    activitiesPromise.then(response => {
+      if (response.data) {
         this.setTableData(response.data);
       }
 
@@ -52,64 +34,60 @@ class FAQ extends Component {
 			console.log(error);
 		});
   }
-	componentWillReceiveProps(nextProps) {
-    var questionsPromise = getAllQuestions();
-    questionsPromise.then(response => {
-      if(response.data){
-        this.setTableData(response.data);
-      }
 
-		}).catch(function (error) {
-			console.log(error);
-		});
-	}
   setTableData(responseData) {
     this.setState({ tableData: responseData })
   }
+
 	onAddRow(row) {
-		if(row && row.question != "" && row.answer != "") {
-			handleInsertQuestion(row);
-    }else{
+		if (row && row.title != "" && row.description!= "" && row.grade != "" && row.station) {
+      handleInsertActivities(row);
+      this.setState((prevState) => {
+        return {tableData: [...prevState.tableData, row]};
+      });
+    } else {
       alert("Please fill out all fields")
     }
-	}
-  onDeleteRow(questionIDs) {
-		if(questionIDs) {
-			handleDeleteQuestions(questionIDs);
+  }
+  
+  onDeleteRow(activityIDs) {
+		if (activityIDs) {
+			handleDeleteActivities(activityIDs);
 		}
   }
+
   beforeSaveCell(row, cellName, cellValue) {
-    if(row.hasOwnProperty("id") && cellValue != "") {
+    if (row.hasOwnProperty("id") && cellValue != "") {
       row[cellName] = cellValue;
-			handleQuestionEdit(row);
-    }else{
+			handleEditActivities(row);
+    } else {
       alert("Please don't leave a field blank");
       return false;
     }
   }
+
   render() {
     if (this.state.tableData.length !== 0) {
-      const inputFieldStyle = {
-        width: "100%"
-      }
-
       const options = {
         onDeleteRow: this.onDeleteRow,
 				onAddRow: this.onAddRow
       };
+
       const cellEdit = {
         mode: 'click', // click cell to edit
         beforeSaveCell: this.beforeSaveCell,
         blurToSave: true
       };
+
       const selectRow = {
         mode: 'checkbox' //radio or checkbox
       };
+
       return (
         <div>
           <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css" />
           <link rel="stylesheet" href="https://npmcdn.com/react-bootstrap-table/dist/react-bootstrap-table-all.min.css"></link>
-          <h3> FAQ </h3>
+          <h3> Activities </h3>
           <BootstrapTable
             data={this.state.tableData}
             insertRow
@@ -123,16 +101,17 @@ class FAQ extends Component {
             options={options}
           >
             <TableHeaderColumn dataField='id' dataSort isKey={ true } hidden hiddenOnInsert>ID</TableHeaderColumn>
-            <TableHeaderColumn dataField='question'>Question</TableHeaderColumn>
-            <TableHeaderColumn dataField='answer'>Answer</TableHeaderColumn>
+            <TableHeaderColumn dataField='title'>Title</TableHeaderColumn>
+            <TableHeaderColumn dataField='description'>Description</TableHeaderColumn>
+            <TableHeaderColumn dataField='station'>Station</TableHeaderColumn>
+            <TableHeaderColumn dataField='grade'>Grade</TableHeaderColumn>
           </BootstrapTable>
         </div>
       );
-    }else{
+    } else {
       return null;
     }
-
   }
 }
 
-export default FAQ;
+export default Activities;
