+++ src/components/Login.jsx
@@ -15,12 +15,12 @@ class Login extends Component {
 				<Grid bsClass="container" style={{"width": "65%"}}>
 					<Row>
             <Col xs={1} md={4}></Col>
-            <Col xs={4} md={4}>
-							<h1 className="text-center login-title" style={{"margin-top": "25px"}}>Sign in to continue to the Dashboard</h1>
+            <Col xs={4} md={6}>
+							<h1 className="text-center login-title" style={{"marginTop": "25px"}}>Sign in to continue to the Dashboard</h1>
 							<div className="account-wall">
                 <Image className="profile-img" src="https://i.imgur.com/jLrP7d1.png" />
 									<form className="form-signin" action="/login" method="post">
-										<FormGroup controlId="formLogin">
+										<FormGroup>
 											<FormControl
 												type="text"
 												name="username"
